Added method to generate cayley table and ease inverse finding
There is now a second method generate_RB_sequence_2, which does the same job as the other class method generate_RB_sequence, but in a more efficient way, by using the Cayley table of the Two qubit Clifford group to find the inverse operation.
Only issue: The Cayley table of 11520*11520 has to be generated once (it is extremely long with the current method _generate_cayley_table() which can be called by simply uncommenting this line in the init function of the class RB_sequence.
To generate the table (only need to do it once), uncomment the line _generate_cayley_table, and run the example code (make sure to set N_sequences to 1 to avoid many iterations!) .

Once the table is generated, it will be saved once and for all in a csv file and we won't need this method anymore.

Another update, a specific version of the scripts for the cross resonance gate is now available.