Simulation option
Added simulatio option to all codes. There was a typos in state tomography and allxy for the strem processing that I fixed. Also de readout macro for state tomography was causing troubles so I fixed as well
forgot to uncomment pi/2

pi/2 missing in ramsey
move back to pi and pi_half

for T1 and Ramsey I moved back to pi and pi_half
Ramsey with frame rotation

addded the modifications so as to Ramsey without update frequency
x270 -> -x90

x270 makes me think as if I'm applying at 3/2*pi_amp pulse because of a 270 degrees rotations, but in the literature the gate needed asks for -x90 which is -pi_amp/2. THis is indeed the definition in the configuration so the name was confusing to me.
black applied

black applied
minor corrections to -x90 -y90

updated waveform and pulses to associate it wiht
changes by testing with QM_QW

test on real device fix typos
Black

black
remove phase rotation


x90 -> pi

move the pulses to `pi` and `pi_half` in the cases that are general 
remove signal detrend in qubit spec


correct labeling to tick 

was missing y180
Simulation option

Added simulatio option to all codes. There was a typos in state tomography and allxy for the strem processing that I fixed. Also de readout macro for state tomography was causing troubles so I fixed as well
forgot to uncomment pi/2

pi/2 missing in ramsey
move back to pi and pi_half

for T1 and Ramsey I moved back to pi and pi_half
Ramsey with frame rotation

addded the modifications so as to Ramsey without update frequency
x270 -> -x90

x270 makes me think as if I'm applying at 3/2*pi_amp pulse because of a 270 degrees rotations, but in the literature the gate needed asks for -x90 which is -pi_amp/2. THis is indeed the definition in the configuration so the name was confusing to me.
black applied

black applied
minor corrections to -x90 -y90

updated waveform and pulses to associate it wiht
changes by testing with QM_QW

test on real device fix typos
Black

black
bring rb from main


Ramsey with frame rotation

instead of frequency detuning
Black