The code needs adaptation
It is for 9 qb 9 resonators, in 2 octave 2 opxs, but easilty can be modified so smaller subet.

I think the correct subset will be 4qb 4 resonator. 

The motivation to have this add hand is that Lincoln labs is distribution candle devices 6qb 6resonators to people to test their fridges and this is something we can reuse in the future.