Added XEB script (#221)
* Added XEB script

This script implements Cross-Entropy Benchmarking which leverages real-time processsing features of the OPX. Namely, the random circuit generation and execution is done on the OPX directly. We also propose tools from Qiskit (should be installed) and Cirq (no need to be installed) for enriching and strengthening the post-processing.

* Adjustment of alignment for two qubit gate

* Modifications following review

Most of the comments have been addressed here.
There is now a dataclass for XEBConfig as suggested, however we leave the user to exploit this structure for saving hyperparameters. Moreover, macros contain both QUA and Python functions used throughout the notebook.

* Applied modifications following second review

* Updated README