Added new class interface for XEB experiment
XEB now follows a similar approach to the 2Q RB Use case, where all the details are abstracted away from the user behind the simple declaration of a class `XEB`. The novelty here is that it is possible to run the experiment from this class interface, but also to simulate it through Qiskit Aer using a method `simulate`, taking as input a Qiskit `AerBackend` object (which is expected to contain a representative noise model of the user's device) and using it to sample from a noisy circuit-level quantum simulation.

All the previous plots that could be retrieved are still available through different plot methods of the `XEBResult`class.