Rearranged config inputs for qubits and qubit pairs
The Config now directly takes the Transmon/TransmonPair QuAM components as input, and the script raises an error if no xy or resonator channel can be found in this qubit. This makes it therefore explicit that it is qubit platform dependent (for ensuring the real-time amplitude matrix modulation component of the script, could be relaxed in the future with proper QuAM abstract components).